Extract SourceTextModuleRecord from CyclicModuleRecord to better match ecmascript
